Move checkbox from src/vs/workbench/browser/positronComponents/positronDynamicModalDialog/components to src/vs/base/browser/ui/positronComponents/checkbox & add PositronList component
