+++ src/vs/base/browser/ui/positronComponents/checkbox/checkbox.css
